The First Ever Build Of GameHub (GameHub 0.1)
Also known as Build M21-20, this build originally released on March 22, 2020. It is in the experimental phase and some bugs have been detected. For example, one bug is that the LIST of games (lst.exe) is missing from the application. If you find more bugs please report them in the comments